Improve sign handling

+++ src/Crypto/Hash/CRC16R_CPP.cpp
@@ -3,8 +3,8 @@
 
 extern "C" void CRC16R_InitTable(UInt16 *tab, UInt16 polynomial)
 {
-	Int32 i = 256;
-	Int32 j;
+	UInt16 i = 256;
+	UInt32 j;
 	UInt16 v;
 	while (i-- > 0)
 	{
